+++ examples/playground/src/App.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import '../../assets/style.css';
-import { onMounted } from 'vue';
 import { vueStore } from './vue';
 import { usePiniaStore } from './pinia';
 import { onKeyDown } from '@vueuse/core';
@@ -9,11 +8,6 @@ import { exit } from '@tauri-apps/plugin-process';
 const piniaStore = usePiniaStore();
 
 onKeyDown('Escape', () => void exit(0));
-
-onMounted(async () => {
-  await piniaStore.$tauri.start();
-  await vueStore.$tauri.start();
-});
 </script>
 
 <template>